phutx-dell

+++ application/models/Category_model.php
@@ -21,21 +21,33 @@ class Category_model extends Base_model {
 
     public function getByName($name) {
         $condition = ['name' => $name];
-        return $this->em->getRepository('Entity\Categories')->findOneBy($condition);        
+        return $this->em->getRepository('Entity\Categories')->findOneBy($condition);
     }
-    
+
     public function delete($slug_name) {
-        $category = $this->getBySlugName($slug_name);
-        $this->em->remove($category);
-        $this->em->flush();
-        
+        if ($this->getBySlugName($slug_name)) {
+            $category = $this->getBySlugName($slug_name);
+            $this->em->remove($category);
+            $this->em->flush();
+            return true;
+        } else {
+            return false;
+        }
     }
-    
-    public function edit(Entity\Categories $category) {
-        $this->em->persist($category);
-        $this->em->flush();
+
+    public function edit($data, $category) {
+        try {
+            $category->setName($data['name']);
+            $category->setSlugName($this->slug->slugify($data['name']));
+            $this->em->persist($category);
+            $this->em->flush();
+            return true;
+        } catch (Exception $e) {
+            log_message('error', $e->getMessage());
+            return false;
+        }
     }
-    
+
     public function save($category) {
         $save = new Entity\Categories();
         $save->setName($category['name']);
@@ -43,4 +55,5 @@ class Category_model extends Base_model {
         $this->em->persist($save);
         $this->em->flush();
     }
+
 }
